Add files via upload
A program that mimics the operations of inside an elevator. More specifically, the program simulates what happens when the user chooses to go to a particular floor and what happens when the user pulls the elevator's fire alarm. Assuming the elevator is in a high-rise building with floors numbered 1 through 10.